% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/madbio.R
\name{madbio}
\alias{madbio}
\title{madbio}
\usage{
madbio(pop, Rt, pE, pI, C, X0, n)
}
\arguments{
\item{pop}{a vector containing the population of each geography}

\item{Rt}{a vector containing the effective reproductive rate of each geography}

\item{pE}{a numeric representing the latent period, or the average number of days to become infected once exposed}

\item{pI}{a numeric representing the recovery period, or the average number of days to recover once infected}

\item{C}{a matrix representing the number of commuters who travel between each pair of regions each day. The origin geographies are represented by the row index, and the destinations are represented by the column index.}

\item{X0}{a matrix representing the number of individuals in each compartment for each geography. Each column represents a geography, and each row represents a different compartment.}

\item{n}{Number of iterations to run}
}
\value{
A data frame object containing the number of individuals in each compartment at each time step for each geography.
}
\description{
test
}
