% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nextIterate.R
\name{nextIterate}
\alias{nextIterate}
\title{nextIterate}
\usage{
nextIterate(pop, C, B, At_const, Xt)
}
\arguments{
\item{pop}{a vector containing the population of each geography}

\item{C}{a matrix representing the number of commuters who travel between each pair of regions each day. The origin geographies are represented by the row index, and the destinations are represented by the column index.}

\item{B}{a vector representing Rt / (pI*pop) which is the reproductive rate per capita, divided by the average recovery time. Calculated in the madbio function.}

\item{At_const}{a matrix representing the change in the number of individuals in each compartment of the model. This is scaled by alpha to get the final At matrix for the linear dynamical system simulation.}

\item{Xt}{a matrix representing the number of individuals in each compartment for each geography. Each column represents a geography, and each row represents a different compartment.}
}
\value{
A data frame object containing the number of individuals in each compartment for each geography at the next time step.
}
\description{
A helper function for the madbio function - usually not used alone. Obtains the next iteration of the m-ADBio model by simulating a single step of the linear dynamical system
}
